Pythoncode + Thesis
A genetic algorithm is a heuristic search method used to find solutions for optimization problems. It is based on the Darwinian theory of evolution. This technique involves the ultimate selection of the fittest (best timetable) from a randomly created population (chromosomes) of solutions for the timetabling problem where each individual (chromosome) represents a timetable. The optimality (perfection) of a chromosome is evaluated by a fitness function based on hard and soft constraints . Genetic algorithms begin by creating a random population of timetables followed by their evaluation according to defined criteria to select parents (timetables) for the next generation which is expected to produce better timetables by way of crossovers and mutations. The process is repeated until a satisfactory solution is reached.

At first I tried using Ant Colony Optimization Algorithm but it was giving an accuracy of 32%, it maybe because of some errors done by me. But I dropped this idea on 28/5/2023. Then afterwards I chose the PYGAD Library on 30/5/2023 but again after giving several tries I couldn’t make the work by myself, In short I needed some in my group to think but not like me which I was all alone. Then finally I decided on 2/6/2023, to start coding(this part took long cause I faced numerous errors and I had to tackle them too) the genetic algorithm by myself but with some help fromresearch papers and various internet resources and side by side as to compare the accuracy of genetic algorithm with another algorithm I chose stimulated annealing as my second algorithm. 

For Local search exploration I used two Functions that I saw in a research paper(Link attached below) that were:
1.	ssn (Simple Searching Neighborhood): Randomly changes the time slot of a class or lab in a given solution.
2.	swn (Swapping Neighborhood): Randomly selects two classes and swaps their time slots.